добавлена таблица заявки и изменены другие

+++ database/migrations/2023_02_04_161703_create_services_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->integer('time');
+            $table->double('time');
             $table->integer('price');
             $table->integer('defect_id');
             $table->timestamps();
